# Conflicts:
#	app/src/main/java/com/example/ddopik/phlogbusiness/ui/MainActivity.java
#	app/src/main/res/values/strings.xml

+++ app/src/main/java/com/example/ddopik/phlogbusiness/ui/userprofile/view/UserProfileActivityView.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserProfileActivityView {
 
     void viewUserProfileLevel(String userLevel); //
-    void viewUserProfileRating(int userRating); //
+    void viewUserProfileRating(float userRating); //
     void viewUserProfileProfileImg(String userImg); //
     void viewUserProfileFullName(String fullName); //
     void viewUserProfileUserName(String userName); //
